Chance cards that go to nearest railroad or utility are slightly broken
I found a bug. It's not a big bug or error,  but it's still a problem. I'll fix this soon. Also. I will soon build a class for feature tests. Alan also found that one of the chance tiles has a community chest script attached. he fixed it for his. I won't fix it for mine because the merge will be easier because we can just use his edit.