Fixed the damage and improved a few small things.
The bidding now runs without any behind the scenes/gameplay bugs as far as I know. I just need to improve the UI (which currently works but looks kinda bad) and the one big thing left for this feature is to add Alan's Scalar class to the UI. I can do this in class pretty easily if Alan shows me how to use his class.